+++ dlib/hash_table/hash_table_kernel_1.h
@@ -145,7 +145,7 @@ namespace dlib
             );
 
             // functions from the enumerable interface
-            inline unsigned long size (
+            inline size_t size (
             ) const;
 
             bool at_start (
