+++ dlib/image_keypoint/binned_vector_feature_image.h
@@ -55,7 +55,7 @@ namespace dlib
             const image_type& img
         );
 
-        inline unsigned long size (
+        inline size_t size (
         ) const;
 
         inline long nr (
@@ -280,7 +280,7 @@ namespace dlib
         typename feature_extractor,
         typename hash_function_type
         >
-    unsigned long binned_vector_feature_image<feature_extractor,hash_function_type>::
+    size_t binned_vector_feature_image<feature_extractor,hash_function_type>::
     size (
     ) const
     {
