+++ dlib/static_set/static_set_kernel_1.h
@@ -116,7 +116,7 @@ namespace dlib
             );
     
             // functions from the enumerable interface
-            inline unsigned long size (
+            inline size_t size (
             ) const;
 
             inline bool at_start (
